Added Operations and made other modifications
Added the Operations cpp/hpp files and skeletons for the two functions truncated and visible.  Did some cosmetic white space elimination on configurationsettings to bring it inline with other files.  Added ConfigurationSettingsPointer to configurationSettings.  Added comments to ID.  Added default constructor to ID (not sure if needed?).    Finished a couple functions in PA.  Uncommented validate for transitionfunction, not sure if it's right, needs testing.